third_party/blink/web_tests/fast/events/show-context-menu-on-mouse-up-expected.txt

setShowContextMenuOnMouseUp setting = false
PASS eventLog is "mousedown contextmenu mouseup"


setShowContextMenuOnMouseUp setting = true
PASS eventLog is "mousedown mouseup contextmenu"


PASS successfullyParsed is true

TEST COMPLETE
